refactor(shopping): extract SemanticKernel categorizer into Yumney.Shopping.Extraction
Mirrors the existing Yumney.Recipes.Extraction / Yumney.MealPlan.Extraction
projects: each module's Semantic Kernel + LLM wiring lives in its own
*.Extraction project so Yumney.Shopping.Infrastructure stays focused on
EF persistence + read-model projections.

What moved out of Yumney.Shopping.Infrastructure:
- Services/SemanticKernelShoppingItemCategorizer.cs
- Services/SemanticKernelOptions.cs
- Services/StubShoppingItemCategorizer.cs (E2E fallback for the same interface)
- AddShoppingItemCategorizer() helper in
  ShoppingInfrastructureServiceCollectionExtensions

What's new in Yumney.Shopping.Extraction:
- All three service files, repointed to the new namespace
  SmartSolutionsLab.Yumney.Shopping.Extraction.Services
- ShoppingExtractionServiceCollectionExtensions.AddShoppingExtraction(...)
  with the same E2E-vs-production wiring shape Recipes.Extraction uses
- ShoppingExtractionModule (IModule) registered alongside the existing
  ShoppingInfrastructureModule from Yumney.Shopping.Api/Program.cs

Other changes:
- Microsoft.SemanticKernel package reference dropped from
  Yumney.Shopping.Infrastructure.csproj (now lives in Shopping.Extraction)
- Yumney.slnx adds the new project to the /src/Shopping/ folder
- ExternalDependencyIsolationTests now also asserts that Shopping.Domain,
  Shopping.Application, Shopping.Infrastructure, and Shopping.Api have no
  dependency on Microsoft.SemanticKernel — guards future drift.

Pure organizational refactor: no behaviour change, IShoppingItemCategorizer
contract and DI lifecycle are identical.
