+++ src/Yumney.Shopping.Api/Program.cs
@@ -2,6 +2,7 @@ using SmartSolutionsLab.Yumney.Shared.Hosting;
 using SmartSolutionsLab.Yumney.Shared.Modules;
 using SmartSolutionsLab.Yumney.Shopping.Api;
 using SmartSolutionsLab.Yumney.Shopping.Application;
+using SmartSolutionsLab.Yumney.Shopping.Extraction;
 using SmartSolutionsLab.Yumney.Shopping.Infrastructure;
 
 var builder = WebApplication.CreateBuilder(args);
@@ -11,6 +12,7 @@ IModule[] modules =
 	new ShoppingApiModule(),
 	new ShoppingApplicationModule(),
 	new ShoppingInfrastructureModule(),
+	new ShoppingExtractionModule(),
 ];
 builder.RegisterServices(modules);
