+++ src/Yumney.Shopping.Extraction/ShoppingExtractionModule.cs
@@ -1,0 +1,13 @@
+using Microsoft.Extensions.Hosting;
+using SmartSolutionsLab.Yumney.Shared.Modules;
+
+namespace SmartSolutionsLab.Yumney.Shopping.Extraction;
+
+public sealed class ShoppingExtractionModule : IModule
+{
+	public IHostApplicationBuilder RegisterServices(IHostApplicationBuilder builder)
+	{
+		builder.Services.AddShoppingExtraction(builder.Configuration);
+		return builder;
+	}
+}
