+++ src/Yumney.Shopping.Infrastructure/ShoppingInfrastructureServiceCollectionExtensions.cs
@@ -1,8 +1,6 @@
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.SemanticKernel;
 using SmartSolutionsLab.Yumney.Recipes.Client;
-using SmartSolutionsLab.Yumney.Shared.Common;
 using SmartSolutionsLab.Yumney.Shared.Events;
 using SmartSolutionsLab.Yumney.Shared.Events.Wolverine;
 using SmartSolutionsLab.Yumney.Shared.Persistence;
@@ -13,7 +11,6 @@ using SmartSolutionsLab.Yumney.Shopping.Infrastructure.ExternalServices;
 using SmartSolutionsLab.Yumney.Shopping.Infrastructure.Persistence;
 using SmartSolutionsLab.Yumney.Shopping.Infrastructure.Persistence.EventStore;
 using SmartSolutionsLab.Yumney.Shopping.Infrastructure.Persistence.ReadModel;
-using SmartSolutionsLab.Yumney.Shopping.Infrastructure.Services;
 using SmartSolutionsLab.Yumney.Users.Client;
 
 namespace SmartSolutionsLab.Yumney.Shopping.Infrastructure;
@@ -48,46 +45,6 @@ public static class ShoppingInfrastructureServiceCollectionExtensions
 		services.AddScoped<IShoppingUserDataPurger, ShoppingUserDataPurger>();
 		services.AddHealthChecks().AddDbContextCheck<ShoppingDbContext>("shoppingdb");
 
-		AddShoppingItemCategorizer(services, configuration);
-
 		return services;
-	}
-
-	private static void AddShoppingItemCategorizer(IServiceCollection services, IConfiguration configuration)
-	{
-		if (configuration.GetValue<bool>("E2ETests"))
-		{
-			services.AddSingleton<IShoppingItemCategorizer, StubShoppingItemCategorizer>();
-			return;
-		}
-
-		services.AddScoped<IShoppingItemCategorizer, SemanticKernelShoppingItemCategorizer>();
-
-		var skOptions = configuration.GetSection(SemanticKernelOptions.SectionName).Get<SemanticKernelOptions>()
-			?? new SemanticKernelOptions();
-		var (provider, modelId, endpoint, apiKey) = skOptions;
-
-		var kernelBuilder = services.AddKernel();
-		switch (provider)
-		{
-			case SemanticKernelOptions.ProviderAzureOpenAI:
-				kernelBuilder.AddAzureOpenAIChatCompletion(modelId, endpoint, apiKey);
-				break;
-			case SemanticKernelOptions.ProviderOllama:
-				var ollama = endpoint.HasValue() ? endpoint : configuration.GetConnectionString("ollama");
-				if (ollama is not null)
-				{
-					kernelBuilder.AddOpenAIChatCompletion(modelId, new Uri(ollama), apiKey: null);
-				}
-
-				break;
-			default:
-				if (apiKey.HasValue())
-				{
-					kernelBuilder.AddOpenAIChatCompletion(modelId, apiKey);
-				}
-
-				break;
-		}
 	}
 }
