+++ tests/Yumney.Architecture.Tests/ExternalDependencyIsolationTests.cs
@@ -11,6 +11,10 @@ public class ExternalDependencyIsolationTests
 	[InlineData("Yumney.Recipes.Domain")]
 	[InlineData("Yumney.Recipes.Application")]
 	[InlineData("Yumney.Recipes.Api")]
+	[InlineData("Yumney.Shopping.Domain")]
+	[InlineData("Yumney.Shopping.Application")]
+	[InlineData("Yumney.Shopping.Infrastructure")]
+	[InlineData("Yumney.Shopping.Api")]
 	public void Layer_ShouldNotDependOn_SemanticKernel(string assemblyName)
 	{
 		var assembly = Assembly.Load(assemblyName);
@@ -22,7 +26,8 @@ public class ExternalDependencyIsolationTests
 
 		result.IsSuccessful.Should().BeTrue(
 			$"{assemblyName} must route LLM calls through abstractions in Application; " +
-			"Semantic Kernel lives in Yumney.Recipes.Extraction only.");
+			"Semantic Kernel lives in the per-module *.Extraction project only " +
+			"(Yumney.Recipes.Extraction, Yumney.Shopping.Extraction).");
 	}
 
 	[Theory]
